reorganize views
