+++ src/components/Main/Main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom'
-import Home from '../../views/Home.jsx';
-import About from '../../views/About.jsx';
+import Home from './views/Home.jsx';
+import About from './views/About.jsx';
 
 const Main = () => {
   return (
