+++ src/components/Main/views/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../components/Button/Button.jsx';
+import Button from '../../Button/Button.jsx';
 
 const Home = () => {
   return (
